requirements.txt
django==4.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-debug-toolbar==4.0.0
django-cors-headers==4.0.0
requests==2.30.0
boto3==1.26.138